Move getArgumentValue to a separate file
